move project feedback form to a component
